requirements.txt
pyautogen
streamlit
pydantic